Class moved into seperate package, methods no longer stuck in a loop.
