+++ src/main/java/com/infoshareacademy/App.java
@@ -1,11 +1,7 @@
 package com.infoshareacademy;
 
-import com.infoshareacademy.api.HolidaysEditor;
+import com.infoshareacademy.edit.HolidaysEditor;
 import com.infoshareacademy.api.HolidaysJsonData;
-import com.infoshareacademy.api.Printer;
-import com.infoshareacademy.menu.MainMenu;
-import com.infoshareacademy.menu.MenuSearch;
-import com.infoshareacademy.search.SearchingInApi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
